refactor: 위치 정보 서비스 리팩토링 dto 삭제제
- LocationInfoService를 인터페이스로 변경하고, 구현체인 LocationInfoServiceImpl을 추가하여 서비스의 구조를 개선
- LocationSearchDto 클래스를 삭제하여 불필요한 DTO를 정리
- 기존의 위치 검색 로직을 LocationInfoServiceImpl로 이동하여 서비스의 책임을 명확히 함
